+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - Allow customizing NFS mount options via config file
 - Updated builds to use Go 1.8
 - Allow ONTAP to split a clone from its parent upon creation
-- Improve efficiency of ONTAP LUN ID selection
+- Improved efficiency of ONTAP LUN ID selection
+- Improved efficiency of ONTAP volume list
 
 ## 17.04.0